Address review: Move deprecation warning to __init__.py
