--- Lib/sqlite3/deprecated.py
@@ -1,10 +1,0 @@
-def __getattr__(name):
-    if name == "OptimizedUnicode":
-        import warnings
-        msg = ("""
-            OptimizedUnicode is obsolete. You can safely remove it from your
-            code, as it defaults to 'str' anyway.
-        """)
-        warnings.warn(msg, DeprecationWarning, stacklevel=2)
-        return str
-    raise AttributeError(f"module 'sqlite3' has no attribute '{name}'")
